+++ package.json
@@ -3,49 +3,43 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@ckeditor/ckeditor5-build-classic": "^37.1.0",
+    "@ckeditor/ckeditor5-react": "^6.0.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.3.4",
     "bootstrap": "^5.2.3",
+    "chart.js": "^4.2.1",
     "firebase": "^9.22.1",
     "formik": "^2.2.9",
+    "history": "^5.3.0",
     "json-server": "^0.17.3",
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.4",
+    "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.2.0",
     "react-loader-spinner": "^5.3.4",
     "react-paginate": "^8.2.0",
     "react-redux": "^8.0.5",
     "react-router": "^6.11.1",
     "react-router-dom": "^6.10.0",
+    "react-script": "^2.0.5",
     "react-scripts": "^5.0.1",
-
     "react-slick": "^0.29.0",
-    "react-toastify": "^9.1.2",
-    "slick-carousel": "^1.8.1",
     "react-social-login-buttons": "^3.9.1",
+    "react-to-print": "^2.14.12",
     "react-toastify": "^9.1.2",
     "reactjs-social-login": "^2.6.2",
-    "redux": "^4.2.1",
-    "uuid": "^9.0.0",
-    
-    "react-script": "^2.0.5",
-    "react-to-print": "^2.14.12",
-    
+    "reactstrap": "^9.1.9",
     "redux": "^4.2.1",
     "redux-thunk": "^2.4.2",
+    "slick-carousel": "^1.8.1",
     "sweetalert2": "^11.7.3",
-    "history": "^5.3.0",
     "toast": "^0.5.4",
-
+    "uuid": "^9.0.0",
     "web-vitals": "^2.1.4",
-    "reactstrap": "^9.1.9",
-    "yup": "^1.1.1",
-    "chart.js": "^4.2.1",
-    "react-chartjs-2": "^5.2.0",
-    "@ckeditor/ckeditor5-build-classic": "^37.1.0",
-    "@ckeditor/ckeditor5-react": "^6.0.0"
+    "yup": "^1.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
