+++ src/components/discount/DiscountFormCss.css
@@ -1,0 +1,17 @@
+    .normal-font{
+        font-family: 'Roboto', sans-serif;font-size: 16px;
+        font-weight: bold;
+    }
+    .title-font{
+        font-family: 'Roboto', sans-serif;font-size: 24px;
+        font-weight: bold;
+        text-align: center;
+    }
+    * {
+        margin: 0;
+        padding: 0;
+        box-sizing: border-box;
+    }
+    tr{
+        height: 60px;
+    }
